Only output profiling info on console if profiling is enabled
