+++ crates/relayer/src/lib.rs
@@ -36,7 +36,6 @@ pub mod foreign_client;
 pub mod keyring;
 pub mod light_client;
 pub mod link;
-pub mod macros;
 pub mod misbehaviour;
 pub mod object;
 pub mod path;
