+++ crates/relayer/src/util.rs
@@ -7,6 +7,7 @@ pub mod diff;
 pub mod iter;
 pub mod lock;
 pub mod pretty;
+pub mod profiling;
 pub mod queue;
 pub mod retry;
 pub mod stream;
